requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
dateparser==1.2.0
rapidfuzz==3.9.7
